Add files via upload
This is a speed limiter script that limits every single car speed to defined one. The speed definition is made in the NUI file an it can be in MPH or KPH. You can set speed based on the street names . The script includes a speed sign shown over the map. If there is any problem with the position feel free to message me. The script is well optimized.